requirements.txt
python-telegram-bot==22.3
openai>=1.107.1
python-dotenv==1.0.0
requests==2.31.0
apscheduler==3.10.4
pytest
httpx==0.27.2
python-telegram-bot-calendar
tenacity==8.2.3
aiosqlite>=0.21.0